Split consoleLog tests into internal and e2e
