+++ packages/typegpu/tests/tgsl/consoleLog.test.ts
@@ -1,18 +1,8 @@
-import { beforeEach, describe, expect, vi } from 'vitest';
-import { namespace } from '../../src/core/resolve/namespace.ts';
-import tgpu, { d } from '../../src/index.js';
-import { ResolutionCtxImpl } from '../../src/resolutionCtx.ts';
-import { deserializeAndStringify } from '../../src/tgsl/consoleLog/deserializers.ts';
-import { CodegenState } from '../../src/types.ts';
+import { describe, expect, vi } from 'vitest';
+import tgpu, { d } from 'typegpu';
 import { it } from 'typegpu-testing-utility';
 
 describe('wgslGenerator with console.log', () => {
-  let ctx: ResolutionCtxImpl;
-  beforeEach(() => {
-    ctx = new ResolutionCtxImpl({ namespace: namespace() });
-    ctx.pushMode(new CodegenState());
-  });
-
   it('Parses console.log in a stray function to a comment and warns', () => {
     using consoleWarnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
 
@@ -802,133 +792,5 @@ describe('wgslGenerator with console.log', () => {
         wrappedCallback(id.x, id.y, id.z);
       }"
     `);
-  });
-});
-
-describe('deserializeAndStringify', () => {
-  it('works for string literals', () => {
-    const data = new Uint32Array([]);
-    const logInfo: (string | d.AnyWgslData)[] = ['String literal'];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "String literal",
-      ]
-    `,
-    );
-  });
-
-  it('works for u32', () => {
-    const data = new Uint32Array([123]);
-    const logInfo: (string | d.AnyWgslData)[] = [d.u32];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "123",
-      ]
-    `,
-    );
-  });
-
-  it('works for vec3u', () => {
-    const data = new Uint32Array([1, 2, 3]);
-    const logInfo: (string | d.AnyWgslData)[] = [d.vec3u];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "vec3u(1, 2, 3)",
-      ]
-    `,
-    );
-  });
-
-  it('works for clumped vectors', () => {
-    const data = new Uint32Array([1, 2, 3, 4, 5, 6]); // no alignment
-    const logInfo: (string | d.AnyWgslData)[] = [d.vec3u, d.vec3u];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "vec3u(1, 2, 3)",
-        "vec3u(4, 5, 6)",
-      ]
-    `,
-    );
-  });
-
-  it('works for multiple arguments', () => {
-    const data = new Uint32Array([1, 2, 3, 456]);
-    const logInfo: (string | d.AnyWgslData)[] = ['GID:', d.vec3u, 'Result:', d.u32];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "GID:",
-        "vec3u(1, 2, 3)",
-        "Result:",
-        "456",
-      ]
-    `,
-    );
-  });
-
-  it('works for arrays', () => {
-    const data = new Uint32Array([1, 2, 3, 4]);
-    const logInfo: (string | d.AnyWgslData)[] = [d.arrayOf(d.u32, 4)];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "[1, 2, 3, 4]",
-      ]
-    `,
-    );
-  });
-
-  it('works for nested arrays', () => {
-    const data = new Uint32Array([1, 2, 3, 4]);
-    const logInfo: (string | d.AnyWgslData)[] = [d.arrayOf(d.arrayOf(d.u32, 2), 2)];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "[[1, 2], [3, 4]]",
-      ]
-    `,
-    );
-  });
-
-  it('works for structs', () => {
-    const data = new Uint32Array([1, 2, 3, 4]);
-    const logInfo: (string | d.AnyWgslData)[] = [d.struct({ vec: d.vec3u, num: d.u32 })];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "{ vec: vec3u(1, 2, 3), num: 4 }",
-      ]
-    `,
-    );
-  });
-
-  it('works for nested structs', () => {
-    const data = new Uint32Array([1, 2, 3, 4, 1]);
-    const logInfo: (string | d.AnyWgslData)[] = [
-      d.struct({
-        nested: d.struct({ vec: d.vec3u, num: d.u32 }),
-        bool: d.bool,
-      }),
-    ];
-
-    expect(deserializeAndStringify(data, logInfo)).toMatchInlineSnapshot(
-      `
-      [
-        "{ nested: { vec: vec3u(1, 2, 3), num: 4 }, bool: true }",
-      ]
-    `,
-    );
   });
 });
